requirements.txt
opencv-python>=4.8.0
Pillow>=9.0.0
numpy>=1.21.0
torch>=1.12.0
lama-cleaner>=1.0.0
onnxruntime>=1.15.0
moviepy>=1.0.3
fastapi>=0.95.0
uvicorn>=0.22.0
pytest>=7.0.0
black>=23.0.0
isort>=5.0.0